Move some things around, delete unused files
